requirements.txt
flask
markdown
pygments  # 代码高亮支持
pyyaml
minio
python-dotenv
flask_sqlalchemy
pymysql